WP serving API and FE from one port
